fix: fix custom wallet startup logic for console wallet (#5429)
Description
---
Added logic so that the wallet will not ask to connect to a detected
base node if a custom base node has already been set in the database.

Motivation and Context
---
The custom-configured base node was ignored.

How Has This Been Tested?
---
System-level testing

What process can a PR reviewer use to test or verify this change?
---
- Start console wallet without `config.wallet.custom_base_node` set
where a base node is running locally and accept the prompt to connect to
the detected base node.
- With _Network_ tab configure a custom base node that is different to
the one that is running locally.
- Exit.
- Restart the console walle.
- Verify that the user is not prompted again to connect to the local
base node.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->