test: feature and scenario names in temp dir path (#5431)
Description
---
Add the Feature name, and Scenario name, to the temp dir structure for
all processes.

Motivation and Context
---
Just to help with debugging.

How Has This Been Tested?
---
Locally

What process can a PR reviewer use to test or verify this change?
---
See CI pass

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->