feat (wip): add sparse merkle tree
Sparse Merkle trees

A sparse Merkle tree is a Merkle tree where the non-empty nodes are stored in a map with the key being the hash
of the node and the value being the node itself.

Merkle trees are a mutable Merkle-ish tree structure, that supports full CRUD operations.

Inclusion _and_ exclusion proofs are supported, and are succinct.
(TODO!)

This implementation assumed that there is a key-value store elsewhere. Only the tree hashes themselves are
handled in this data structure.

When constructing a new tree, a hashing algorithm is specified. This is used to hash the non-leaf nodes. The
"values" provided to the tree must already be a hash, and should have been generated from a different hashing
algorithm to the one driving the tree, in order to prevent second pre-image attacks.
delete works now.