feat: ui for template registration in console wallet (#5444)
Description
---
Added code template registration tab to the console wallet.
The UI shows proper errors. It will not erase the fields when something
is wrong.
Autofill everything that it can from the binary url (template name,
version, type, repo link).
Computes the hash for the template.

Motivation and Context
---
The autofill works like this:
If the link is e.g.
`https://github.com/tari-project/tari/blob/development/template.12.wasm`.
It will fill the repo as `https://github.com/tari-project/tari`
Template name `template`, template version `12`, template type `WASM:1`.
The hash is always computed as of now, because there is no way for the
user to generate it manually.

How Has This Been Tested?
---
Manually.

What process can a PR reviewer use to test or verify this change?
---
Register a code template, and look into the VN UI.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: Andrejs Gubarevs <1062334+agubarev@users.noreply.github.com>
Co-authored-by: stringhandler <mikethetike@tari.com>