fix(nextnet): updates the block to the in use on stagenet now (#5455)
Description
---
Sets the genesis block time to now and not a future data.
Release a new version

Motivation and Context
---
We cannot mine on stagenet till the genesisblock timestamp has past.

How Has This Been Tested?
---

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify