ci(fix): for binaries build, fix metadata issues (#5473)
      Description
Resolve an issue where the metadata generation is not functioning

Motivation and Context
Make the CI more green

How Has This Been Tested?
Built in local fork

What process can a PR reviewer use to test or verify this change?
At the end of the binaries build, check for errors in the summary of
workflow

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->