feat!: remove witness mr (#5476)
Description
---
This removes the witness mr from the header and use of the witness data
hash. Combines it into a single output hash.

Motivation and Context
---
We use two hashes for each output, the output hash and the witness hash.
The output hash excludes the metadata signature and range proof. 
This is stored separately as the witness data in a separate Merkle
mountain range.
This complicates matters and there is no use for having two separate
hashes.

This makes it much clearer in having a single hash for each output.
Rather than requiring two hashes for mutability per output.


How Has This Been Tested?
---
Passes all unit tests

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->
New genesis block and new wallet db