refactor: various fixes for clarity (#5475)
Description
---
A few fixes found while going through the `core/consensus` crate.

Motivation and Context
---
Some names were ambiguous and some needed more detail 

How Has This Been Tested?
---
With clippy, haven't tested manually, but I'm still going through it

What process can a PR reviewer use to test or verify this change?
---
Most of these are comments or renames, so if you agree with the new
comments or names, then that should be sufficient

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->