ci(fix): build only bins and include in archive/packages (#5479)
Description
Limit build to only binaries that are included in the archives and
packages

Motivation and Context
In the name of efficiency, which save about 30% of build time if we only
build the binaries packaged

How Has This Been Tested?
Built in local fork and used archive to run a local node, wallet and
miner.

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---
Removed tari_mining_helper_ffi from archives and packages

- [ ] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [x] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->