fix: coinbase recovery (#5487)
Description
---
Fixes the coinbase recovery. 
Speeds up recovery searching by looking on the private key.

Motivation and Context
---
The current code can recover the private key of the coinbase, but it
searches in the `CommitmentMask` till index 1_000_000 not finding a
match. Then it imports the key. For coinbase keys, the keys are derived
not from `CommitmentMask` but from `Coinbase` so it much searches that
index.

This bug will also slow down the wallet spending transactions by
blocking the `CommitmentMask` branch via the mutex for that massive
index search. This is now removed and will speed up the wallet also.

How Has This Been Tested?
---
Unit tests, manual

What process can a PR reviewer use to test or verify this change?
---