fix(core/keymanager): use tokio rwlock for keymanagers (#5494)
Description
---
Replaces futures-rs Mutex with tokio's RwLock for key manager locks

Motivation and Context
---
Allow multiple concurrent reads for keymanagers. In general, we prefer
to use tokio's locks with the tokio runtime (no investigation done into
performance differences).

If there was a particular reason to use futures-rs Mutex we can close
this.

How Has This Been Tested?
---
Existing tests, this change is not benchmarked.

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->