chore(core/chain_storage): remove outdated annotation and comment (#5496)
Description
---
Removed outdated `too_many_lines` annotation and comment, it was a
leftover from before the function was refactored.

Motivation and Context
---
Code cleanup as result of the `chain_storage` internal audit 

How Has This Been Tested?
---
Testing does not apply

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>