chore: update local coverage script (#5532)
The source coverage scripts are incredibly outdated. This updates to
work (though I'm not convinced it's 100% accurate - function % results
seems very wrong). But that's for another PR

## Deleted scripts

A few of the utility scripts have been deleted since they don't appear
to be used anywhere or have any real use cases.
If this is not the case, comment below!


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->