fix(console-wallet): fix possible subtract underflow panic in list (#5535)
Description
---
Fix possible subtract with underflow panic in console wallet list state

Motivation and Context
---
Ref #5465 

Unable to reproduce this exact panic manually however added a test case
which causes the same panic. I was previously able to reproduce this
every time, either this has been fixed in other changes or the
reproduction steps did not include some mix of circumstances that were
pertinent to the bug. In any case, this PR fixes a bug in select_first
function which uses the binary compliment where a boolean inequality
check was likely intended !self.num_items == 0 vs self.num_items != 0,
though strangely the correct logic is 'if self.num_items == 0 {'. On an
empty list this causes a non-existent first item at index 0 to be
selected, a call to update_list_state would panic.

How Has This Been Tested?
---
Unit test that reproduces the same underflow panic in #5465, although
the send to contact list does not use select_first
Manually

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->