fix: update code coverage approach
Changing out the grcov approach with cargo-llvm-cov which is supposed to
be much quicker.

Allows us to easily run code coverage from command line:

```
./scripts/source_coverage.sh
```

or in CI.

We can easily manage the crates that are included by editing the
`ignored_crates` array.

In this PR only a small subset of crates are included. This is to check
that everything is working. A follow up will add more crates to the
coverage tests.