fix: update code coverage approach (#5540)
Changing out the grcov approach with cargo-llvm-cov which is supposed to
be much quicker.

Allows us to easily run code coverage from command line:

```
./scripts/source_coverage.sh
```

or in CI.

We can easily manage the crates that are included by editing the
`ignored_crates` array.

In this PR only a small subset of crates are included. This is to check
that everything is working. A follow up will add more crates to the
coverage tests.


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->