fix: remove optional timestamp verification bypass (#5552)
Description
---
In previous refactors we had to bypass timestamp checks during header
sync as it we where asking the database for the header timestamps. We
now cache the timestamps when syncing negating this limitation.
We can thus remove this optional check and always check the header
timestamps pass the median rule

How Has This Been Tested?
---
Pass all unit tests
Manual sync of multiple networks.