feat: add a not before proof (#5560)
Description
---
Added a `not before proof` to the genesis block. The data to be added is
just a simple Vec<u8>, which can represent anything, but must be limited
in size.

_**Note:** This is to implement what some members of the Tari community
have been discussing. The `not before proof` or how it is implemented is
not my idea; I merely executed it. That being said, I had some fun with
choosing the examples._

Motivation and Context
---
The basic idea behind adding a `not before proof` to the genesis block
is to provide some level of authenticity. If we would, for example,
insert a digitized headline of the _New York Times_ just before mainnet
launches, it will provide a timestamped reference to a widely recognized
and verifiable event and prove that the blockchain has not been in
existence before that time.

How Has This Been Tested?
---
Genesis block sanity test

What process can a PR reviewer use to test or verify this change?
---
Review changes

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

Co-authored-by: SW van Heerden <swvheerden@gmail.com>