Remove statement from sparse Merkle tree proofs
refactor: rework proof data structures

The SMT proof data structures have been reworked. There is now a type
for In- and ExclusionProofs. The proofs are created with the respective
`::from_tree(...)` methods.

Both types have a `validate` method that lets a verifier validate the
proof. For inclusion proofs, the verifier provides the expected key,
value and root hash. For exclusion proofs, they provide the expcted key
and root hash.

This implementation also returns a `NonviableProof` error when trying to
create an inclusion proof on a key that does not exist, or vice-versa
for exclusion proofs. This is probably more in line with expected
behaviour.

Implementation notes:

There is a fair degree of overlaps between the two types, so we
introduce a trait, `MerkleProofDigest` that captures the common
behaviour.

Some unused functions in bit_utils have been removed.

We introduce `PathIterator`, that converts a key into a iterator of
traversal directions. This is DoubleEndedIterator, so it can be reversed
etc. Side note, implementing DEI has some nasty gotchas (like you must
override the default nth_back, len and size_hint methods, or else
connector iters like take and rev just don't work). This was surprising
and took a few hours to debug.

docs: add docstrings to proof structures