ci(fix): manually download checksums tool for windows (#5581)
Description
Download PowerShell tools that are used directly out of github vs using
```choco upgrade psutils```, which is outdated and not working.

Motivation and Context
Fixes Windows checksum

How Has This Been Tested?
Build in local fork

What process can a PR reviewer use to test or verify this change?
Check that Windows build is successful with checksums

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify