feat: chat ffi status callback (#5583)
Description
---
This PR adds a callback manager and the ability to pass a callback into
the chat ffi for use when a contact's status changes.

Motivation and Context
---
Requested by the mobile team.
Pattern similar to wallet ffi callbacks, although purposefully not
copied, it just seemed to be the natural way to go 😄

How Has This Been Tested?
---
Added integration test.
Although we get 16 callback calls. Not _totally_ unexpected but more
than I was... expecting.

What process can a PR reviewer use to test or verify this change?
---
Read some code.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: SW van Heerden <swvheerden@gmail.com>