ci(fix): update Windows installer for Minotari (#5614)
Description
Correct some of the paths for Minotari

Motivation and Context
Fix the Windows installer build

How Has This Been Tested?
Builds in local fork.

What process can a PR reviewer use to test or verify this change?
Check that a Windows installers is successfully created.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>